requirements.txt
albumentations==1.3.0
einops==0.8.0
erqa==1.1.2
ffms2==0.4.5.5
grpcio==1.66.1
h5py==3.11.0
imageio==2.28.1
kornia==0.7.0
lightning-utilities==0.11.8
lit==16.0.3
lpips==0.1.4
numpy==1.24.3
opencv-python==4.8.0.74
opencv-python-headless==4.7.0.72
pandas==2.0.1
Pillow==9.5.0
pytorch-lightning==1.9.0
PyYAML==6.0
scikit-image==0.20.0
scikit-learn==1.2.2
scikit-video==1.1.11
scipy
timm==1.0.11
torch==2.0.1
torch-tb-profiler==0.4.3
torchmetrics==1.5.1
torchvision==0.15.2
tqdm==4.66.1

